rename: 오류 확인 및 절대 경로 변경(#206)
