+++ src/features/game/types/boxscore.ts
@@ -80,6 +80,14 @@ export interface Pitcher {
   tb: string;
   w: number;
   wls: string;
+  era?: string;
+  start?: number;
+  sv?: number;
+  hold?: number;
+  wra?: string;
+  innDisplay?: string;
+  bb?: number;
+  hp?: number;
 }
 
 export interface ScheduleInfo {
