+++ src/features/player/components/detail/PlayerRecordChart.tsx
@@ -1,14 +1,13 @@
-import { useState } from 'react';
-
 import {
   ChartLabelList,
   CustomBarChart,
   CustomLineChart,
 } from '@/features/common';
 import { SubTitle } from '@/features/common';
+import { RecentRecord, YearRecord } from '@/features/player/types/detail';
 import { cn } from '@/lib/utils';
+import { useState } from 'react';
 import Skeleton from 'react-loading-skeleton';
-import { RecentRecord, YearRecord } from '../../types/detail';
 import { RecordTableAccordion } from './RecordTableAccordion';
 
 export interface Config {
