+++ src/features/player/components/detail/RecentRecordTab.tsx
@@ -4,8 +4,8 @@ import {
   recentPitcherConfig,
 } from '@/constants/chart-config';
 import { SubTabsTrigger } from '@/features/common';
+import { usePlayer } from '@/features/player/hooks/usePlayer';
 import { useParams } from 'react-router';
-import { usePlayer } from '../../hooks/usePlayer';
 import { PlayerRecordChart } from './PlayerRecordChart';
 
 function RecentRecordTab() {
